Don't allow transitive scala_library dependencies onto the classpath
Without this patch, scala and java "deps of deps" end up on the
classpath. This isn't the way java_library rules work:

    A rule X can access the code in Y if there exists a dependency
    path between them that begins with a deps edge followed by zero
    or more exports edges.

I verified my understanding of this behavior with a small test repo,
copying one of the tests from this repo and changing 'scala_' to
'java_':
https://gist.github.com/colinmarc/c683652ef9a7d817c5e7ce9ca1236a6d

This patch fixes that, and all the tests that were incorrectly
depending on the behavior. It also adds some negative tests for the
behavior, under `test_transitive_deps`.

Unfortunately, this is going to be majorly backwards incompatible
with existing scala repositories.