Add _jdk dependency to necessary places (#172)
The default @local_jdk just imports _java and _javac as raw binaries
without establishing a dependency on/runfiles of _jdk. Thus, all uses of
those should manually add _jdk as a dependency.

In the scalac execution ation that compiles scala, also added _jdk in
case _scalac ever was replaced by a binary that wasn't a java_binary.

_jar_bin does not need a _jdk dependency since is itself a java_binary

Note that earlier versions of this PR had a change to use ijars. This
was reverted (although the comment referring to using ijar amended to
the proper line) to keep this PR as a purely bugfix PR that should
strictly increase the chance something compiles