Adds the _java_runtime attribute to fix toolchain use in Scala.
When executing non-locally, the wrong version of Java would be used,
referencing a path that didn't exist. This change adds the
_java_runtime attribute to have to locate the java runtime on the
machine where the Scala code is being executed, and selects the
correct configuration for the host_javabase.

As a minor cleanup, this also removes the now no-longer-necessary
_java attribute.

TESTED=Ran unit tests and also ran a custom Scala unit test through:
previously, the Scala unit test failed with an incorrect java path
error, now it passes.

Change-Id: I9c2829bded04ec2d9d014094bd9e72cd984fe6e6