Regression test for data attr in scala_test_suite
Modifies the TestSuitePassesKwArgs test-case to test that the data
attribute is passed through by scala_test_suite. This is a stronger
test-case since it doesn't just test that attributes like data,
jvm_flags, and scalac_jvm_flags are accepted by scala_test_suite. But,
it also tests that data and jvm_flags are actually passed through to
scala_test. The test-case tries to read from the file defined in these
attributes.