Remove manually forwarded arguments
These are now forwarded as part of the kwargs argument.
Additionally, the data argument is now forwarded to scala_test_suite,
instead of being quietly ignored as it was before.
This addresses the following comment
https://github.com/bazelbuild/rules_scala/pull/627#issuecomment-432085769

An additional advantage is that attribute defaults don't have to be kept
in sync between the *_suite and non-suite rules/macros.

closes #627