Forward kwargs in scala_library_suite and scala_test_suite (#622)
* Forward **kwargs

From `scala_library_suite` to `scala_library`, and
from `scala_test_suite` to `scala_test`.

closes #621

* Add regression test for #621

* Disable linting

See https://github.com/bazelbuild/rules_scala/pull/622#issuecomment-428662365

* Remove manually forwarded arguments

These are now forwarded as part of the kwargs argument.
Additionally, the data argument is now forwarded to scala_test_suite,
instead of being quietly ignored as it was before.
This addresses the following comment
https://github.com/bazelbuild/rules_scala/pull/627#issuecomment-432085769

An additional advantage is that attribute defaults don't have to be kept
in sync between the *_suite and non-suite rules/macros.

closes #627

* Regression test for data attr in scala_test_suite

Modifies the TestSuitePassesKwArgs test-case to test that the data
attribute is passed through by scala_test_suite. This is a stronger
test-case since it doesn't just test that attributes like data,
jvm_flags, and scalac_jvm_flags are accepted by scala_test_suite. But,
it also tests that data and jvm_flags are actually passed through to
scala_test. The test-case tries to read from the file defined in these
attributes.

* Rename test targets

Addressing the following comment
https://github.com/bazelbuild/rules_scala/pull/622#pullrequestreview-167213724