Fix transitive thrift symbol resolution in IntelliJ
Currently, given the following build file structure:

scrooge_scala_library -> thrift_library -> thrift_library

where "A -> B" means "A depends on B"

Only symbols in the leftmost thrift_library will resolve in IJ.
This is because when the compile scrooge jars are gathered up,
transitive jars are not gathered. This fixes this behavor by
adding transitive jars to the ScroogeAspectInfo.output_files
depset.