Fix transitive thrift symbol resolution in IntelliJ (#660)
* Fix transitive thrift symbol resolution in IntelliJ

Currently, given the following build file structure:

scrooge_scala_library -> thrift_library -> thrift_library

where "A -> B" means "A depends on B"

Only symbols in the leftmost thrift_library will resolve in IJ.
This is because when the compile scrooge jars are gathered up,
transitive jars are not gathered. This fixes this behavor by
adding transitive jars to the ScroogeAspectInfo.output_files
depset.

* Add transitive outputs field to ScroogeAspectInfo. Add dep attr to scrooge rule.

* Use new style providers in twitter_scrooge

This addresses https://github.com/bazelbuild/rules_scala/issues/564

tl;dr: In order to support thrift resolution in IJ, we had to use old-style
providers to return the metadata in a shape that IJ understood. Since bazel
0.17, java_common.merge has been changed to support using new-style providers.

This uses new-style providers and gets rid of the previous hack, but note
that this means this change *isn't compatible with bazel < 0.17*

* Remove unnecessary code

It appears that this mostly just works given the new behavior of java_common.merge (outputs are merged), so I was able to remove all the code that was gathering up outputs manually.

* Add unit tests for scrooge rule outputs

* Bump to latest version of skylib

* Run tests with bazel 0.17.1

* Run CI tests with upper and lower bounds.