Fix root path for generated files (#688)
The owner's workspace_root for proto files that are generated by a
genrule is empty. This means that the root path ends with a trailing
slash. PBGenerateRequest expects the root path to not end in a slash
and removes an additional character (see line 17) which is the expected
path separator after the workspace_root component of the path.
For generated files, this means that the first character of the proto
file path is cut off, and thus the scalapb compiler cannot read the
file.
Instead of removing and additional character, using stripPrefix provides
a less fragile way of removing a leading slash.