Include toolchains attrs in example
Turns out I made some mistakes on the native toolchain name vs. the implementtion name. Not for the first time.

Also had to add TemplateVariableInfo to the core scala_toolchain as required when using the standard toolchains attr.
Simply initialized it with an empty map (not really familiar with this provider ...)