JacocoInstrumenter jar reproducibility (#1040)
* Test for JacocoInstrumenter jar reproducibility

* Make JacocoInstrumenter jar file stable by creating the jar file with JarCreator

When creating a jar file with FileSystems.newFileSystem() it adds timestamps
of files to the jar, that makes those jar files non-reproducible.
We can use JarCreator (with a temp directory) to create a stable jar.

* Resolving review comments

* JacocoInstrumenter: change Paths.get to path.resolve

* Refactor deleteTempDir to DirectoryUtils and unit test it

* Force failing unit test

* Revert "Force failing unit test"

This reverts commit e1577eb95e9bb52be8a47b63f1185e83cb1649b2.

* Reproduce "File name too long" exception with new JacocoInstrumenter implementation

* JacocoInstrumenter: fix "File name too long" exception for complicated Scala constructs

For very long class names the previous implementation failed.
Write to the jar file directly (as was the case with FileSystems) when
possible and create temporary files with short names if this is not possible.

Create metadata directory (instead of temp dir) the same way as rules_java
does.

* Fixing review comments

Removing unused imports.
setNormalize(true) is the default in JarCreator.
Comment fix.