WIP - Create rule for a `scala_service_provider`
A `scala_service_provider` is like a `scala_library`, excepts it exports
a list of implementations for a number of services in its output jar.

Use the phases mechanism to wrap around the `scala_library` rule with
minimal boilerplate.

This is a work-in-progress and I would like to ask a few questions
before submitting it as a PR:

* Does it make sense to create a new target for a
`scala_service_provider`?
* If so, would it be a welcome addition to `rules_scala` or should we
simply keep this implementation inside our own code base?
* Is it right to use the phases mechanism as I did to add a phase for
writing services to the output jar? Is there a more idiomatic way to
achieve low duplication between this new rule and `scala_library`?
* Does the name: `scala_service_provider` sound good? It is in line with
the "SPI" terminology but may conflict with other concepts in the Bazel
world. Would `scala_spi_plugin` be more explicit/suitable?
* Is there a better way to implement the test than a shell script under
shell? Maybe using the `test_sh` target and doing it in a build file?