Add executable as a direct dependency
This fixes a bug introduced by
https://github.com/bazelbuild/rules_scala/pull/958. Before #958,
`rootpath` on a Scala binary resolved to the wrapper binary so other
rules could treat it as an opaque executable. After that PR `rootpath`
now resolves to the JAR instead of the wrapper binary.

Adding the executable back as a direct dependency fixes that.