Reintroduce implicit_compile_deps
Reverts #801

ScalaPB features like extra imports and type mappers don't work with the current implementation of scala_proto_library.
This adds back `implicit_compile_deps` tha allows the user to add extra implicit dependencies added to all scala_proto_library targets.

Reverted some of the changes in #801. This should work with the latest versions of Bazel since toolchain transitions have been introduced.

This is not the ideal solution since changes to theses dependencies will trigger a rebuild of all the proto targets.

But it can be an escape hatch for people already using these features and are looking to migrate to Bazel.