Extended proto rule
## What?
Adds a new rule for compiling protobuf to scala. This rule is not aspect based and that allows the user to add extra dependencies to the compilation.

## Why?
The current proto rule does not support adding extra dependencies becaue it's aspect based.

## How?
Extracted some of the functionality of the aspect based rule and reused it in the new rule.