JacocoRunner script: update for Bazel 5.0+
Upgraded jacoco package name for Bazel 5.0+.

Upgraded jacoco from 0.8.3 to 0.8.6.

Jacoco upgrade to 0.8.6 was made in Bazel in:
https://github.com/bazelbuild/bazel/issues/11674
https://github.com/bazelbuild/bazel/commit/cb7c1a2ec8de92828a018e2a3450a06f1d1c7efa

Removed options for workarounds where those have been fixed in the meantime:

Bazel's handling for branch coverage was fixed in:
https://github.com/bazelbuild/bazel/issues/12696

Instead of changing the existing script for building Jacoco, added a new
one that can be used for Bazel 5.0.