Fix long file names in scalac action when handling srcjars
Resolves https://github.com/bazelbuild/rules_scala/issues/1322

Instead of using full path of srcjar as a destination directory
prepend index to srcjar file name.
Index corresponds to order in which file was passed to scalac.

Also considered computing hash of full path, but full path
contains platform specific labels ie darwin, linux etc
so not sure that would be a good approach.