JacocoRunner script: update for Bazel 6.0+
Added a parameter to specify the Bazel major version (to have a single
script for Bazel 5 and 6).
Manually tested on a repo with both Bazel versions.

Had to add a new patch file for Bazel 6.
Added new Bazel branch with the Jacoco interface implementation, updated
for Bazel 6.3.2.

Updated manual test_coverage_with_local_jacocorunner to run the script
with Bazel 6 provided as a parameter (it was failing after the repo's
upgrade to Bazel 6).