Don't output unused dependency warning and errors when compilation fails. (#1548)
When a scala source is malformed, the AST parsing dependency checking step comes
back with no/very little dependencies used. It then prints out a bunch of these
warnings which are not necessarily true. As a workaround, let's not print the
unused dependency warnings if the build has already failed before hand.