Bind toolchains to Scala version
Use `target_settings` to indicate which Scala version is supported by the toolchain.
This will become useful when multiple toolchains for each Scala version are defined.

The choice of toolchains to cover by this (first) migration is quite arbitrary,
but should be enough to enable the basic cross-build use cases.

Co-authored-by: mkuta <mkuta@virtuslab.com>