Build against latest bazel version instead of last_green
Reverts https://github.com/bazelbuild/rules_scala/pull/1349

This check should be optional however it fails all builds because bazel info fails

Error is related to workspace setup

Since we are in process of moving to bzlmod I think we should not bother with that now.

Especially when reproducing locally it gives different error.