Restore `EXCLUDED_ARTIFACTS` mechanism
The new mechanism is more thorough and robust than the version from
"Make scala-parser-combinators a root artifact". It will not only
exclude artifacts from the newly generated dict, but will remove `deps`
labels for newly excluded artifacts from the existing dict.

There are no artifacts in the `EXCLUDED_ARTIFACTS` set yet, but we'll
add `com.google.guava:listenablefuture` to it once we get to updating
gRPC and Guava. I wanted this mechanism to stand on its own for clarity.