Extract `scala_proto_artifact_ids`
This will support the upcoming schema whereby `scala_toolchains` will
import all artifact IDs via macros like this, and invoke `repositories`
itself. This will avoid instantiating the same artifact repository
multiple times for different frameworks that depend on it.

Under `WORKSPACE`, this isn't a problem, but it's an error under Bzlmod,
as a module extension can only instantiate a repository once. The commit
message to toolchainize Scalafmt will provide more details on this.