Vary the specs2 version by the Scala version
Different versions of specs2 support different Scala versions, so it's
necessary to use a different version, depending on which Scala version
we're on. v4.10.6 is the last to support Scala 2.11, v4.20.9 is the last
to support Scala 2, and the latest version as of the time this commit
was written, v5.5.8, only supports Scala 3.