Add Java 24 compatibility
The SecurityManager is a hollow shell that just throws exceptions in
Java 24. We should avoid calling it. This will make tests that call
System.exit crash the worker.

Bazel is doing the same thing for now, see
https://github.com/bazelbuild/bazel/issues/24354

It is often possible to edit code to avoid calling System.exit, and guarding
against it requires more effort in Java 24+, likely involving attaching
an agent. Until someone actually has that need, it doesn't seem worth doing.