Replace `scala_deps.toolchains` with tag classes
Replaces the `scala_deps.toolchains` tag class, which contained boolean
parameters for selecting toolchains, with a series of individual tag
classes for each toolchain.

Suggested by @simuons in the following comment as an alternative
implementation that could allow for adding attributes to each
toolchain's tag class.

- https://github.com/bazelbuild/rules_scala/pull/1722#discussion_r2051570874

We already had tag classes for `scalafmt`, `scala_proto`, and
`twitter_scrooge` options. Now the presence of these tag classes, along
with the tag classes added in this commit, translates to a `True`
argument for `scala_toolchains()`.